requirements.txt
asgiref==3.4.1
Django==3.2
django-cors-headers==3.10.1
djangorestframework==3.13.1
Pillow==8.4.0
psycopg2-binary==2.9.2
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
